Create Hello World
as you can see I am a new starter to coding. For some reason this code is not working and I don't know why. I have re typed this a few times. I am following from a coding book and I am using TextWrangler on a Mac. 

There appears to be a problem with the <!DOCTYPE html> line (I think)

any suggestions welcome.